Dynamicall adjust x and y offset depending on space constraints

+++ forge/src/main/java/com/teamderpy/shouldersurfing/forge/ClientEventHandler.java
@@ -54,7 +54,7 @@ public class ClientEventHandler
 	@SuppressWarnings("resource")
 	public static void computeCameraAnglesEvent(ComputeCameraAngles event)
 	{
-		ShoulderRenderer.getInstance().offsetCamera(event.getCamera(), Minecraft.getInstance().level, event.getPartialTick());
+		ShoulderRenderer.getInstance().offsetCamera(event.getCamera(), Minecraft.getInstance().level, (float) event.getPartialTick());
 	}
 	
 	@SubscribeEvent
